Deleted Multimedia context and relocate Video to Acccounts one
