+++ lib/rumbl/accounts/accounts.ex
@@ -4,7 +4,7 @@ defmodule Rumbl.Accounts do
     The Accounts context.
   """
 
-  alias Rumbl.Accounts.User
+  alias Rumbl.Accounts.{User, Video}
   alias Rumbl.Repo
 
   def list_users() do
@@ -31,6 +31,33 @@ defmodule Rumbl.Accounts do
 
   def remove_user(%User{} = item) do
     Repo.delete(item)
+  end
+
+  # Start Videos
+  def list_videos do
+    Repo.all(Video)
+  end
+
+  def get_video!(id), do: Repo.get!(Video, id)
+
+  def create_video(attrs \\ %{}) do
+    %Video{}
+    |> Video.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  def update_video(%Video{} = video, attrs) do
+    video
+    |> Video.changeset(attrs)
+    |> Repo.update()
+  end
+
+  def delete_video(%Video{} = video) do
+    Repo.delete(video)
+  end
+
+  def change_video(%Video{} = video, attrs \\ %{}) do
+    Video.changeset(video, attrs)
   end
 
 end
