+++ lib/rumbl_web/controllers/video_controller.ex
@@ -1,21 +1,21 @@
 defmodule RumblWeb.VideoController do
   use RumblWeb, :controller
 
-  alias Rumbl.Multimedia
-  alias Rumbl.Multimedia.Video
+  alias Rumbl.Accounts
+  alias Rumbl.Accounts.Video
 
   def index(conn, _params) do
-    videos = Multimedia.list_videos()
+    videos = Accounts.list_videos()
     render(conn, "index.html", videos: videos)
   end
 
   def new(conn, _params) do
-    changeset = Multimedia.change_video(%Video{})
+    changeset = Accounts.change_video(%Video{})
     render(conn, "new.html", changeset: changeset)
   end
 
   def create(conn, %{"video" => video_params}) do
-    case Multimedia.create_video(video_params) do
+    case Accounts.create_video(video_params) do
       {:ok, video} ->
         conn
         |> put_flash(:info, "Video created successfully.")
@@ -27,20 +27,20 @@ defmodule RumblWeb.VideoController do
   end
 
   def show(conn, %{"id" => id}) do
-    video = Multimedia.get_video!(id)
+    video = Accounts.get_video!(id)
     render(conn, "show.html", video: video)
   end
 
   def edit(conn, %{"id" => id}) do
-    video = Multimedia.get_video!(id)
-    changeset = Multimedia.change_video(video)
+    video = Accounts.get_video!(id)
+    changeset = Accounts.change_video(video)
     render(conn, "edit.html", video: video, changeset: changeset)
   end
 
   def update(conn, %{"id" => id, "video" => video_params}) do
-    video = Multimedia.get_video!(id)
+    video = Accounts.get_video!(id)
 
-    case Multimedia.update_video(video, video_params) do
+    case Accounts.update_video(video, video_params) do
       {:ok, video} ->
         conn
         |> put_flash(:info, "Video updated successfully.")
@@ -52,8 +52,8 @@ defmodule RumblWeb.VideoController do
   end
 
   def delete(conn, %{"id" => id}) do
-    video = Multimedia.get_video!(id)
-    {:ok, _video} = Multimedia.delete_video(video)
+    video = Accounts.get_video!(id)
+    {:ok, _video} = Accounts.delete_video(video)
 
     conn
     |> put_flash(:info, "Video deleted successfully.")
